requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amadeus==8.0.0
amqp==5.1.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dataclasses-json==0.5.7
decorator==5.1.1
Django==4.2
django-bootstrap-v5==1.0.11
django-celery-results==2.5.1
django-environ==0.10.0
environ==1.0
executing==1.2.0
ezgraphics @ file:///Users/harshpatel/Downloads/ezgraphics-2.2.tar.gz
frozenlist==1.3.3
gptcache==0.1.21
idna==3.4
ipython==8.13.1
jedi==0.18.2
Jinja2==3.1.2
kombu==5.2.4
langchain==0.0.142
llama-index==0.5.27
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.23.0
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
regex==2023.5.2
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.2.2
tiktoken==0.3.3
tk==0.1.0
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
